HSEARCH-3263 Remove implementation code specific to ES versions <5.6
