+++ backend/elasticsearch/src/main/java/org/hibernate/search/backend/elasticsearch/document/model/impl/esnative/IndexType.java
@@ -17,21 +17,6 @@ package org.hibernate.search.backend.elasticsearch.document.model.impl.esnative;
  */
 public enum IndexType {
 
-	/**
-	 * @deprecated ES2 only.
-	 */
-	@Deprecated
-	NOT_ANALYZED,
-	/**
-	 * @deprecated ES2 only.
-	 */
-	@Deprecated
-	ANALYZED,
-	/**
-	 * @deprecated ES2 only.
-	 */
-	@Deprecated
-	NO,
 	TRUE,
 	FALSE
 	;
