--- legacy/elasticsearch/src/main/java/org/hibernate/search/elasticsearch/analyzer/impl/Elasticsearch2Analyzers.java
@@ -1,15 +1,0 @@
-/*
- * Hibernate Search, full-text search for your domain model
- *
- * License: GNU Lesser General Public License (LGPL), version 2.1 or later
- * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
- */
-package org.hibernate.search.elasticsearch.analyzer.impl;
-
-
-/**
- * @author Yoann Rodiere
- */
-public class Elasticsearch2Analyzers {
-
-}
