+++ legacy/elasticsearch/src/main/java/org/hibernate/search/elasticsearch/schema/impl/model/DataType.java
@@ -24,12 +24,6 @@ public enum DataType {
 			return true;
 		}
 	},
-	/**
-	 * @deprecated Only used in Elasticsearch 2.x. Use TEXT or KEYWORD instead.
-	 */
-	@SerializedName("string")
-	@Deprecated
-	STRING,
 	@SerializedName("text")
 	TEXT,
 	@SerializedName("keyword")
